+++ packages/expo-router/src/LocationProvider.tsx
@@ -28,6 +28,7 @@ export function getRouteInfoFromState(
   };
 }
 
+// Checks whether the deepest focused route is an index route.
 function isIndexPath(state: State) {
   const route = state.routes[state.index ?? state.routes.length - 1];
   if (!route) {
@@ -36,19 +37,7 @@ function isIndexPath(state: State) {
     return isIndexPath(route.state);
   }
 
-  // Index routes on the same level as a layout do not have `index` in their name
-  if (route.params && 'screen' in route.params) {
-    return route.params.screen === 'index';
-  }
-
-  // The `params` key will not exist if there are no params
-  // So we need to do a positive lookahead to check if the route ends with /index
-  // Nested routes that are hoisted will have a name ending with /index
-  // e.g name could be /user/[id]/index
   if (route.name.match(/.+\/index$/)) return true;
-
-  // The state will either have params (because there are multiple _layout) or it will be hoisted with a name
-  // If we don't match the above cases, then it's not an index route
 
   return false;
 }
@@ -77,14 +66,13 @@ export function getNormalizedStatePath(
   };
 }
 
+// TODO: Consolidate parameter decoding across the router.
 function decodeParams(params: Record<string, string>) {
   const parsed: Record<string, any> = {};
 
   for (const [key, value] of Object.entries(params)) {
     try {
-      if (key === 'params' && typeof value === 'object') {
-        parsed[key] = decodeParams(value);
-      } else if (Array.isArray(value)) {
+      if (Array.isArray(value)) {
         parsed[key] = value.map((v) => decodeURIComponent(v));
       } else {
         parsed[key] = decodeURIComponent(value);